skills/earnings-analysis-skill/requirements.txt
pandas>=1.5.0
matplotlib>=3.6.0
plotly>=5.11.0
requests>=2.28.0
beautifulsoup4>=4.11.0
tabula-py>=2.5.0
PyPDF2>=3.0.0
kaleido>=0.2.1
